requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.1
ffmpeg-python>=0.2.0
sentencepiece>=0.1.99
sacremoses>=0.0.53
rouge-score>=0.1.2
nltk>=3.8.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0
jiwer>=3.0.3
evaluate>=0.4.0
pandas>=2.0.0
huggingface-hub>=0.19.0
matplotlib>=3.7.0
seaborn>=0.12.0
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.25.0